Move numpy functions into numpy
