+++ seedbank/__init__.py
@@ -15,7 +15,6 @@ import numpy as np
 
 from seedbank._keys import RNGKey, SeedLike, make_seed
 from seedbank._state import SeedState
-from seedbank.numpy import NPRNGSource
 
 try:
     __version__ = version("seedbank")
@@ -27,6 +26,7 @@ _log = logging.getLogger(__name__)
 _root_state = SeedState()
 
 __all__ = [
+    "make_seed",
     "initialize",
     "derive_seed",
     "numpy_rng",
@@ -178,63 +178,5 @@ def int_seed(
         return seed.generate_state(words)
 
 
-def numpy_rng(
-    spec: Optional[NPRNGSource] = None,
-) -> np.random.Generator:
-    """
-    Get a NumPy random number generator.  This is similar to
-    :func:`sklearn.utils.check_random_state`, but it returns a
-    :class:`numpy.random.Generator` instead.
-
-    Args:
-        spec:
-            The spec for this RNG. Can be any of the following types:
-
-            * :data:`SeedLike`
-            * :class:`numpy.random.RandomState`
-            * :class:`numpy.random.Generator`
-
-    Returns:
-        A random number generator.
-    """
-
-    if isinstance(spec, np.random.Generator):
-        return spec
-    elif isinstance(spec, np.random.RandomState):
-        return np.random.Generator(spec._bit_generator)
-    elif spec is None:
-        return np.random.default_rng(derive_seed())
-    else:
-        seed = make_seed(spec)
-        return np.random.default_rng(seed)
-
-
-def numpy_random_state(spec: Optional[NPRNGSource] = None) -> np.random.RandomState:
-    """
-    Get a legacy NumPy random number generator (:class:`numpy.random.mtrand.RandomState`).
-    This is similar to :func:`sklearn.utils.check_random_state`.
-
-    Args:
-        spec:
-            The spec for this RNG.  Can be any of the following types:
-
-            * :data:`SeedLike`
-            * :class:`numpy.random.RandomState`
-            * :class:`numpy.random.Generator`
-
-    Returns:
-        A random number generator.
-    """
-
-    if isinstance(spec, np.random.RandomState):
-        return spec
-    elif isinstance(spec, np.random.Generator):
-        return np.random.RandomState(spec.bit_generator)
-    elif spec is None:
-        return np.random.RandomState(int_seed(seed=derive_seed()))
-    else:
-        seed = make_seed(spec)
-        return np.random.RandomState(int_seed(seed=seed))
-
-
 from seedbank.cupy import cupy_rng  # noqa: E402
+from seedbank.numpy import numpy_random_state, numpy_rng  # noqa: E402
